refactor: rename repository files to use plural naming convention
